requirements.txt
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.1
Django==2.1.5
django-cors-headers==3.0.2
django-cors-middleware==1.2.0
django-extensions==1.7.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.2
idna==2.8
Markdown==3.0.1
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
Unipath==1.1
urllib3==1.24.1
